requirements.txt
# MCP Server dependencies
fastmcp>=0.1.0
mcp>=1.0.0

# Agent dependencies
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0

# Web framework
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.0.0

# GitHub integration
httpx>=0.26.0
aiohttp>=3.9.0

# Utilities
python-dotenv>=1.0.0
rich>=13.7.0
typer>=0.9.0
